Wip make thumnail a default include in v2 api
