+++ catalogue_api/api/src/main/scala/uk/ac/wellcome/platform/api/requests/Requests.scala
@@ -3,7 +3,7 @@ package uk.ac.wellcome.platform.api.requests
 import com.twitter.finagle.http.Request
 import com.twitter.finatra.request.{QueryParam, RouteParam}
 import com.twitter.finatra.validation.{Max, Min}
-import uk.ac.wellcome.display.models.WorksIncludes
+import uk.ac.wellcome.display.models.{V1WorksIncludes, V2WorksIncludes, WorksIncludes}
 
 sealed trait ApiRequest {
   val request: Request
@@ -21,7 +21,7 @@ trait MultipleResultsRequest extends ApiRequest {
 case class V1MultipleResultsRequest(
   @Min(1) @QueryParam page: Int = 1,
   @Min(1) @Max(100) @QueryParam pageSize: Option[Int],
-  @QueryParam includes: Option[WorksIncludes],
+  @QueryParam includes: Option[V1WorksIncludes],
   @QueryParam query: Option[String],
   @QueryParam _index: Option[String],
   request: Request
@@ -32,7 +32,7 @@ case class V1MultipleResultsRequest(
 case class V2MultipleResultsRequest(
   @Min(1) @QueryParam page: Int = 1,
   @Min(1) @Max(100) @QueryParam pageSize: Option[Int],
-  @QueryParam include: Option[WorksIncludes],
+  @QueryParam include: Option[V2WorksIncludes],
   @QueryParam query: Option[String],
   @QueryParam _index: Option[String],
   request: Request
@@ -47,7 +47,7 @@ trait SingleWorkRequest {
 
 case class V1SingleWorkRequest(
   @RouteParam id: String,
-  @QueryParam includes: Option[WorksIncludes],
+  @QueryParam includes: Option[V1WorksIncludes],
   @QueryParam _index: Option[String],
   request: Request
 ) extends SingleWorkRequest {
@@ -56,7 +56,7 @@ case class V1SingleWorkRequest(
 
 case class V2SingleWorkRequest(
   @RouteParam id: String,
-  @QueryParam include: Option[WorksIncludes],
+  @QueryParam include: Option[V2WorksIncludes],
   @QueryParam _index: Option[String],
   request: Request
 ) extends SingleWorkRequest
