+++ catalogue_api/api/src/test/scala/uk/ac/wellcome/platform/api/works/v2/ApiV2WorksTest.scala
@@ -83,37 +83,6 @@ class ApiV2WorksTest extends ApiV2WorksTestBase {
     }
   }
 
-  it("renders the items if the items include is present") {
-    withV2Api {
-      case (apiPrefix, _, indexNameV2, itemType, server: EmbeddedHttpServer) =>
-        val work = createIdentifiedWorkWith(
-          items = createIdentifiedItems(count = 1) :+ createUnidentifiableItemWith()
-        )
-
-        insertIntoElasticsearch(indexNameV2, itemType, work)
-
-        eventually {
-          server.httpGet(
-            path = s"/$apiPrefix/works/${work.canonicalId}?include=items",
-            andExpect = Status.Ok,
-            withJsonBody = s"""
-               |{
-               | "@context": "https://localhost:8888/$apiPrefix/context.json",
-               | "type": "Work",
-               | "id": "${work.canonicalId}",
-               | "title": "${work.title}",
-               | "contributors": [ ],
-               | "items": [ ${items(work.items)} ],
-               | "subjects": [ ],
-               | "genres": [ ],
-               | "production": [ ]
-               |}
-          """.stripMargin
-          )
-        }
-    }
-  }
-
   it(
     "returns the requested page of results when requested with page & pageSize") {
     withV2Api {
@@ -262,91 +231,6 @@ class ApiV2WorksTest extends ApiV2WorksTestBase {
     }
   }
 
-  it(
-    "includes a list of identifiers on a list endpoint if we pass ?include=identifiers") {
-    withV2Api {
-      case (apiPrefix, _, indexNameV2, itemType, server: EmbeddedHttpServer) =>
-        val works = createIdentifiedWorks(count = 2).sortBy { _.canonicalId }
-
-        val identifier0 = createSourceIdentifier
-        val identifier1 = createSourceIdentifier
-
-        val work0 = works(0).copy(otherIdentifiers = List(identifier0))
-        val work1 = works(1).copy(otherIdentifiers = List(identifier1))
-
-        insertIntoElasticsearch(indexNameV2, itemType, work0, work1)
-
-        eventually {
-          server.httpGet(
-            path = s"/$apiPrefix/works?include=identifiers",
-            andExpect = Status.Ok,
-            withJsonBody = s"""
-               |{
-               |  ${resultList(apiPrefix, totalResults = 2)},
-               |  "results": [
-               |   {
-               |     "type": "Work",
-               |     "id": "${work0.canonicalId}",
-               |     "title": "${work0.title}",
-               |     "contributors": [ ],
-               |     "identifiers": [ ${identifier(work0.sourceIdentifier)}, ${identifier(
-                                identifier0)} ],
-               |     "subjects": [ ],
-               |     "genres": [ ],
-               |     "production": [ ]
-               |   },
-               |   {
-               |     "type": "Work",
-               |     "id": "${work1.canonicalId}",
-               |     "title": "${work1.title}",
-               |     "contributors": [ ],
-               |     "identifiers": [ ${identifier(work1.sourceIdentifier)}, ${identifier(
-                                identifier1)} ],
-               |     "subjects": [ ],
-               |     "genres": [ ],
-               |     "production": [ ]
-               |   }
-               |  ]
-               |}
-          """.stripMargin
-          )
-        }
-    }
-  }
-
-  it(
-    "includes a list of identifiers on a single work endpoint if we pass ?include=identifiers") {
-    withV2Api {
-      case (apiPrefix, _, indexNameV2, itemType, server: EmbeddedHttpServer) =>
-        val otherIdentifier = createSourceIdentifier
-        val work = createIdentifiedWorkWith(
-          otherIdentifiers = List(otherIdentifier)
-        )
-        insertIntoElasticsearch(indexNameV2, itemType, work)
-
-        eventually {
-          server.httpGet(
-            path = s"/$apiPrefix/works/${work.canonicalId}?include=identifiers",
-            andExpect = Status.Ok,
-            withJsonBody = s"""
-               |{
-               | "@context": "https://localhost:8888/$apiPrefix/context.json",
-               | "type": "Work",
-               | "id": "${work.canonicalId}",
-               | "title": "${work.title}",
-               | "contributors": [ ],
-               | "identifiers": [ ${identifier(work.sourceIdentifier)}, ${identifier(
-                                otherIdentifier)} ],
-               | "subjects": [ ],
-               | "genres": [ ],
-               | "production": [ ]
-               |}
-          """.stripMargin
-          )
-        }
-    }
-  }
-
   it("searches different indices with the ?_index query parameter") {
     withV2Api {
       case (apiPrefix, _, indexNameV2, itemType, server: EmbeddedHttpServer) =>
@@ -469,7 +353,7 @@ class ApiV2WorksTest extends ApiV2WorksTestBase {
   }
 
   it(
-    "includes the thumbnail field if available and we use the thumbnail include") {
+    "shows the thumbnail field if available") {
     withV2Api {
       case (apiPrefix, _, indexNameV2, itemType, server: EmbeddedHttpServer) =>
         val work = createIdentifiedWorkWith(
@@ -484,7 +368,7 @@ class ApiV2WorksTest extends ApiV2WorksTestBase {
 
         eventually {
           server.httpGet(
-            path = s"/$apiPrefix/works?include=thumbnail",
+            path = s"/$apiPrefix/works",
             andExpect = Status.Ok,
             withJsonBody = s"""
                |{
